Corrected capitalization on Sham69 track name
Updated the capitalization Sham 69's "If The Kids Are United" track on
match conventional capitalization and, more importantly, match the
track name in the punk-bands-completed.html file that tutorial users
will be comparing against.

Also added the "lang" attribute to the html tags on both files to
appease the Nu Html Checker.